refactor: move deposit logic down
